+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix typehints
 * Update Filepond JS dependencies
 * Add jfif to the common extension for image/jpeg (fix test)
+* Update psalm (fixes Override)
 
 ## 1.7.1